+++ src/components/form-message.tsx
@@ -1,4 +1,4 @@
-import { EXTERNAL_LINKS } from "@/lib/links";
+import { EXTERNAL_LINKS } from "@/lib/constants/external-links";
 import clsx from "clsx";
 import { CheckCircle, Info, XCircle } from "lucide-react";
