+++ src/features/missions/components/mission-details.tsx
@@ -5,7 +5,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { DifficultyBadge } from "@/components/ui/difficulty-badge";
 import { MissionIcon } from "@/components/ui/mission-icon";
 import { YouTubeSubscribeButton } from "@/features/mission-detail/components/youtube-subscribe-button";
-import { YOUTUBE_MISSION_CONFIG } from "@/lib/constants";
+import { YOUTUBE_MISSION_CONFIG } from "@/lib/constants/mission-config";
 import { dateFormatter } from "@/lib/formatter";
 import type { Tables } from "@/lib/types/supabase";
