+++ src/lib/constants/external-links.ts
